requirements.txt
Flask>=3.0.0
requests>=2.31.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pytz>=2023.3
openpyxl>=3.1.2
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
psycopg[binary]>=3.1.0
